requirements.txt
appdirs==1.4.3
cffi==1.13.2
Click==7.0
cryptography==2.8
distlib==0.3.0
dnspython==1.16.0
dominate==2.4.0
filelock==3.0.12
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
Flask-LDAP==0.1.6
Flask-Limiter==1.1.0
Flask-Login==0.5.0
Flask-WTF==0.14.2
future==0.18.2
importlib-metadata==1.5.0
itsdangerous==1.1.0
Jinja2==2.10.3
limits==1.4.1
MarkupSafe==1.1.1
mod-wsgi==4.7.0
mysql-connector-python==8.0.19
protobuf==3.6.1
pyad==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
pyOpenSSL==19.1.0
python-ldap==3.2.0
pywin32==227
six==1.14.0
virtualenv==20.0.1
visitor==0.1.3
Werkzeug==0.16.0
wfastcgi==3.0.0
WTForms==2.2.1
zipp==2.2.0
